# Conflicts:
#	src/main/resources/de/uniks/stpmon/team_m/lang/lang_zh.properties

+++ src/main/java/de/uniks/stpmon/team_m/dto/MonsterAttributes.java
@@ -1,7 +1,7 @@
 package de.uniks.stpmon.team_m.dto;
 
 public record MonsterAttributes(
-        int health,
+        float health,
         int attack,
         int defense,
         int speed
